Fix some problems with the latest raspbian image

+++ stealthrpi-packages.sh
@@ -1,19 +1,25 @@
 #!/bin/bash -e
 
-MESA_ARMHF_VERSION="19.1.0~git20190429.9628631a380-1~rpt3"
+MESA_ARMHF_VERSION="19.2.0~rc1-1~bpo10+1~rpt3"
 MESA_ARM64_VERSION="18.3.6-2"
 
-PACKAGES="$(apt list --installed |& grep '+rpi' | sed 's#/.*$##')"
+DRM_ARMHF_VERSION="2.4.99-1~bpo10~1"
+DRM_ARM64_VERSION="2.4.97-1"
+
+PACKAGES="$(apt list --installed |& grep '+rpi\|'${MESA_ARMHF_VERSION}'\|'${DRM_ARMHF_VERSION} | sed 's#/.*$##')"
 PACKAGE_COUNT=$(wc -w <<< ${PACKAGES})
 if [ "${PACKAGE_COUNT}" == "$(ls *+stealthrpi*.deb 2> /dev/null | wc -w)" ]; then
     echo "All stealthrpi packages appear to be here"
     echo "rm *.deb if you wish to regenerate"
     exit 0
 fi
-echo "Fetching +rpi armhf packages: ${PACKAGES//$'\n'/ }"
+echo "Fetching ${PACKAGE_COUNT} +rpi and mesa armhf packages: ${PACKAGES//$'\n'/ }"
 apt-get download ${PACKAGES}
 
-for DEB in *+rpi*.deb; do
+for DEB in *.deb; do
+    if [[ ${DEB} == *"+stealthrpi"* ]]; then
+        continue
+    fi
     PKGNAME="${DEB/_*}"
     rm -rf tmp; mkdir -p tmp; cd tmp
     ar xf ../${DEB}
@@ -22,7 +28,15 @@ for DEB in *+rpi*.deb; do
     if grep -q "${MESA_ARMHF_VERSION}" control; then
         if ! grep -q "${MESA_ARM64_VERSION}" control; then
             sed -i "s/${MESA_ARMHF_VERSION}/${MESA_ARM64_VERSION}/g" control
-	fi
+        fi
+    fi
+    if grep -q "${DRM_ARMHF_VERSION}" control; then
+        if ! grep -q "${DRM_ARM64_VERSION}" control; then
+            sed -i "s/${DRM_ARMHF_VERSION}/${DRM_ARM64_VERSION}/g" control
+        fi
+    fi
+    if grep -q "${DRM_ARMHF_VERSION/-*}" control; then
+        sed -i "s/${DRM_ARMHF_VERSION/-*}/${DRM_ARM64_VERSION/-*}/g" control
     fi
     sed -i 's/+rpi[[:digit:]]//g' control
     tar cJf ../control.tar.xz *
@@ -34,6 +48,9 @@ for DEB in *+rpi*.deb; do
     tar cJf ../data.tar.xz *
     cd .. # leave newdata/
     NEW_DEB="${DEB//+rpi/+stealthrpi}"
+    if [[ ${NEW_DEB} != *"+stealthrpi"* ]]; then
+        NEW_DEB="${DEB//_armhf/+stealthrpi_armhf}"
+    fi
     ar cr ../"${NEW_DEB}" debian-binary control.tar.xz data.tar.xz
     rm -f ../"${DEB}"
 
@@ -41,3 +58,4 @@ for DEB in *+rpi*.deb; do
     echo "Generated ${NEW_DEB}"
 done
 echo "Successfully generated ${PACKAGE_COUNT} \"stealthrpi\" packages"
+
